Added persistence (and update) of 'averageHashDuration' in NodeState (via GenericFileCachedUrlSet. Changed default hashtest length from 1 second to 10 seconds in SystemMetrics. Named the TreeWalkThread.
git-svn-id: svn://svn.code.sf.net/p/lockss/svn/lockss-daemon/trunk@1153 4f837ed2-42f5-46e7-a7a5-fa17313484d4

+++ test/src/org/lockss/daemon/TestSystemMetrics.java
@@ -1,5 +1,5 @@
 /*
- * $Id: TestSystemMetrics.java,v 1.5 2003-01-03 00:12:04 aalto Exp $
+ * $Id: TestSystemMetrics.java,v 1.6 2003-04-04 23:50:11 aalto Exp $
  */
 
 /*
@@ -74,7 +74,7 @@ public class TestSystemMetrics extends LockssTestCase {
     assertEquals(estimate, estimate2);
   }
 
-  public static void configHashParams(int duration, int step)
+  public static void configHashParams(long duration, int step)
       throws IOException {
     String s = SystemMetrics.PARAM_HASH_TEST_DURATION + "=" + duration;
     String s2 = SystemMetrics.PARAM_HASH_TEST_BYTE_STEP + "=" + step;
